Feature/s20 file upload download

+++ controllers/admin.js
@@ -1,5 +1,5 @@
-const { validationResult } = require('express-validator/check');
-
+const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 
 /**
@@ -27,11 +27,27 @@ const Product = require('../models/product');
  */
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const description = req.body.description;
     const price = req.body.price;
     //const userId = request.user.id;
+    if (!image) {
+        return res.status(422).render('admin/add-product', {
+          pageTitle: 'Add Product',
+          path: '/admin/add-product',
+          editing: false,
+          hasError: true,
+          product: {
+            title: title,
+            price: price,
+            description: description
+          },
+          errorMessage: 'Attached file is not an image.',
+          validationErrors: []
+        });
+      }
     const errors = validationResult(req);
+
     if (!errors.isEmpty()) {
         console.log(errors.array());
         return res.status(422).render('admin/add-product', {
@@ -41,7 +57,6 @@ exports.postAddProduct = (req, res, next) => {
           hasError: true,
           product: {
             title: title,
-            imageUrl: imageUrl,
             price: price,
             description: description
           },
@@ -49,6 +64,8 @@ exports.postAddProduct = (req, res, next) => {
           validationErrors: errors.array()
         });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title:title, 
@@ -128,7 +145,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const updatedImage = req.file;
     const updatedDesc = req.body.description;
 
     const errors = validationResult(req);
@@ -141,7 +158,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDesc,
                 _id: prodId
@@ -158,8 +174,11 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle;
             product.price = updatedPrice;
-            product.imageUrl = updatedImageUrl;
             product.description = updatedDesc;
+            if (updatedImage){
+                //fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = updatedImage.path;
+            }
             return product.save()
                 .then(result => {
                     console.log('Producto actualizado exitosamente!');
@@ -176,14 +195,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-        console.log('Producto eliminado exitosamente!');
+    Product.findById(prodId)
+      .then(product => {
+        if (!product) {
+          return next(new Error('Product not found.'));
+        }
+        //fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+      })
+      .then(() => {
+        console.log('DESTROYED PRODUCT');
         res.redirect('/admin/products');
-    })
-    .catch(err =>{
+      })
+      .catch(err => {
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
-    });
-};
+      });
+  };
